skip typecheck if app does not have tsconfig
